rename partial pdp components
