+++ packages/extension-chakra-storefront/src/pages/product-detail/partials/product-details-simple.jsx
